+++ platforms/h5/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.34](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.33...@vtj/h5@0.16.34) (2026-05-06)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
+## [0.16.33](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.32...@vtj/h5@0.16.33) (2026-04-29)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
+## [0.16.32](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.31...@vtj/h5@0.16.32) (2026-04-28)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
+## [0.16.31](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.30...@vtj/h5@0.16.31) (2026-04-28)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
+## [0.16.30](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.29...@vtj/h5@0.16.30) (2026-04-28)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
+## [0.16.29](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.28...@vtj/h5@0.16.29) (2026-04-28)
+
+**Note:** Version bump only for package @vtj/h5
+
+
+
+
+
 ## [0.16.28](https://gitee.com/newgateway/vtj/compare/@vtj/h5@0.16.27...@vtj/h5@0.16.28) (2026-04-28)
 
 **Note:** Version bump only for package @vtj/h5